refactor: move ClassThatBindsAGenericType to testclasses
