+++ src/test/java/spoon/test/generics/testclasses3/ClassThatBindsAGenericType.java
@@ -1,10 +1,10 @@
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
 
 import java.io.File;
 import java.util.ArrayList;
 
 @SuppressWarnings("serial")
-class ClassThatBindsAGenericType extends ArrayList<File> {
+public class ClassThatBindsAGenericType extends ArrayList<File> {
 
 	public static void main(String[] args) throws Exception {
 	}
